\name{OptionalInteger-class}
\Rdversion{1.1}
\docType{class}
\alias{OptionalInteger-class}
\alias{OptionalCharacter-class}
\alias{OptionalNumeric-class}

\title{Class \code{"OptionalInteger"}}
\description{
Utility classes to allow for inclusion of "NULL" an a class instance, for memory efficiency. Each one of these is a class union between the underlying class ("integer", "character" and "numeric") and "NULL".
}
\section{Objects from the Class}{These classes are typically part of more complex classes representing outputs of ensemble fit functions.}
\section{Methods}{
No methods defined with class "OptionalInteger" in the signature.
}

\author{
Alireza S. Mahani, Mansour T.A. Sharabiani
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\linkS4class{Regression.FitObj}}, \code{\linkS4class{BaseLearner.CV.FitObj}}, \code{\linkS4class{Regression.CV.FitObj}}
}

\keyword{classes}
