\name{RegressionEstObj-class}
\Rdversion{1.1}
\docType{class}
\alias{RegressionEstObj-class}

\title{Class \code{"RegressionEstObj"}}
\description{
Union of (converted) S3 classes for individual base learners as defined in their corresponding packages. The special class "character" has been added to allow for returning filepaths when saving estimation objects to disk.
}
\section{Objects from the Class}{Objects from this class are typically returned as part of \code{FitObj} family of classes.}
\section{Methods}{
No methods defined with class "RegressionEstObj" in the signature.
}

\author{
Alireza S. Mahani, Mansour T.A. Sharabiani
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\linkS4class{BaseLearner.FitObj}}
}

\keyword{classes}
